Rename components and remove bootstrap classes
